Refactor: rename interface module
