+++ pyiron_workflow/composite.py
@@ -11,7 +11,7 @@ from typing import Literal, Optional, TYPE_CHECKING
 
 from bidict import bidict
 
-from pyiron_workflow.interfaces import Creator, Wrappers
+from pyiron_workflow.create import Creator, Wrappers
 from pyiron_workflow.io import Outputs, Inputs
 from pyiron_workflow.node import Node
 from pyiron_workflow.node_package import NodePackage
