+++ pyiron_workflow/semantics.py
@@ -18,10 +18,11 @@ from typing import Optional
 
 from bidict import bidict
 
+from pyiron_workflow.has_interface_mixins import HasLabel
 from pyiron_workflow.snippets.logger import logger
 
 
-class Semantic(ABC):
+class Semantic(HasLabel, ABC):
     """
     An object with a unique semantic path.
